fix: missing admin gates + tick wrap + Vk image extent (audit batch 5)
Fifth-pass audit. Eight more parallel sub-agents flagged ~70
candidates; these four are the ones that survived line-by-line
verification.

shell: `instr` (disassembles arbitrary kernel memory) and `addr2sym`
(reverse-symbol lookup, exposes KASLR layout / kernel function bases)
were NOT admin-gated. Same threat class as the already-gated
`memdump` and `read` verbs — an unprivileged shell-redirect could
dump kernel code or enumerate the symbol map. Gate both via the
existing RequireAdmin helper.

sched: the loadavg-sample interval check at sched.cpp:1852 used the
straight unsigned form `now_ticks - s_last_sample >= kSamplePeriod`.
Once `g_tick_now` wraps past UINT64_MAX the subtraction produces a
huge value and the gate either fires every tick or never again
(depending on the wrap distance), freezing loadavg sampling. Switch
to the signed-diff form already used by TickReached at line ~993 so
the comparison survives wraparound.

subsystems/graphics/graphics_vk: VkCreateImage stored an unbounded
VkExtent3D into the image record. The companion
VkGetImageMemoryRequirements computed `width*height*4` as the
backing size — with width and height both untrusted u32, the
multiplication can exceed u64, and a future real-GPU bring-up would
attempt to back a multi-PB image. VkGetPhysicalDeviceImageFormat-
Properties already advertises maxExtent={16384,16384,1}; enforce
the same cap up-front at creation so the size computation can't
wrap.

False positives surfaced and dismissed:

  - subsystems/translation NtDoQueryDefaultLocale / UILanguage
    "register confusion": NT-call convention is rdi=syscall#,
    rsi=arg1, rdx=arg2; both handlers read from the correct slot
    for their argument index.
  - linux mmap file off+page*sz "u64 wrap to OOB": the wrapped
    offset must be < file_size (≤4 KiB scratch) to enter the
    copy, and the wrap can't simultaneously bypass that bound and
    point past the scratch — verified by case analysis.
  - vmap commit_mask u32 shift overflow: pages is already capped
    at kWin32VmapRegionPagesMax (32) at line 280; the shift index
    can't reach 32.
  - clock-app stopwatch/timer footer "stack overflow" (re-flagged
    in batch 5): both inner string-copy loops are self-bounded.
  - shell `port` / `kdbg` missing admin gate (re-flagged): both
    ARE admin-gated.
  - TCP seq+u32(payload_len) "overflow": payload_len comes from
    tcp_len-data_off where tcp_len is bounded by the IP frame's
    total_len; the cast can't accept a 4 GB payload.

+++ kernel/sched/sched.cpp
@@ -1849,7 +1849,15 @@ void OnTimerTick(u64 now_ticks)
     {
         static volatile u64 s_last_sample = 0;
         const u64 kSamplePeriod = 5ULL * ::duetos::time::TickHz();
-        if (now_ticks - s_last_sample >= kSamplePeriod)
+        // Signed-diff form: `now_ticks - s_last_sample` is unsigned,
+        // so once `now_ticks` wraps past UINT64_MAX -> 0 the
+        // straight comparison would produce a huge positive value
+        // and immediately satisfy the condition forever, or the
+        // opposite (depending on the wrap distance), freezing
+        // loadavg sampling. Cast through i64 so the comparison
+        // works across the wrap. Mirrors the TickReached pattern
+        // at line ~993.
+        if (static_cast<i64>(now_ticks - s_last_sample) >= static_cast<i64>(kSamplePeriod))
         {
             s_last_sample = now_ticks;
             u32 runnable = 0;
